% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyzeSound}
\alias{analyzeSound}
\title{Analyze sound}
\usage{
analyzeSound(x, samplingRate = NULL, silence = 0.03,
  entropyThreshold = 0.9, windowLength = 50, wn = "gaussian", step = 25,
  zp = 0, zpCep = 2^13, pitch_floor = 75, pitch_ceiling = 3500,
  maxPitchCands = 4, voiced_threshold_autocor = 0.75,
  voiced_threshold_cep = 0.45, voiced_threshold_spec = 0.5,
  specPitchThreshold_nullNA = 0.5, slopeSpec = 0.1, widthSpec = 150,
  pitchSpec_only_peak_weight = 0.51, priorMean = NA, priorSD = NA,
  cutoff_dom = 6000, dom_threshold = 0.1, shortestSyllable = 20,
  shortestBreak = 60, interpolWindow = 3, interpolTolerance = 0.3,
  interpolCert = 0.3, certWeight = 0.5, runSnake = T,
  snakeSmoothingStep = 0.05, snakeIterMultiplier = 2, plotSnake = F,
  smooth = T, smoothIdx = 1, plot = T, savePlot = F,
  savePlotPath = "~/Downloads/temp/", contrast = 1, brightness = 1,
  ylim = c(0, 5), smooth_vars = c("pitch", "dom"))
}
\arguments{
\item{x}{path to a .wav file or a vector of amplitudes with specified
samplingRate}

\item{samplingRate}{sampling rate of \code{x} (only needed if \code{x} is a
numeric vector, rather than a .wav file)}

\item{windowLength}{length of fft window (ms)}
}
\description{
Acoustic analysis of a single sound file.
}
