% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_segment.R
\name{evaluate_params}
\alias{evaluate_params}
\title{Evaluate segmentation quality}
\usage{
evaluate_params(p, pars_to_optimize, pars_bounds, fitness_measure, myfolder,
  key)
}
\arguments{
\item{p}{numeric vector of evaluated values of parameters to optimize}

\item{pars_to_optimize}{the names of parameters to optimize (character
vector)}

\item{pars_bounds}{a list of bounds within which pars are to stay. Anything
outside the bounds is not evaluated (the function returns 1). For ex., if
we optimize \code{smooth_ms} and \code{smooth_overlap}, reasonable
pars_bounds might be list(low = c(5, 0), high = c(500, 95))}

\item{fitness_measure}{are we interested in optimizing the count of syllables
(\code{nSyllables}) or bursts (\code{nBursts})?}

\item{myfolder}{path to where the .wav files live}

\item{key}{a vector containing the "correct" number of syllables in each file}
}
\value{
Returns Pearson's correlation between fitness measure (nBursts or
  nSyllables) and the key.
}
\description{
Internal soundgen function.
}
\details{
Called by \code{\link{optimizeSegment}}.
}
