% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{evaluatePars}
\alias{evaluatePars}
\title{Evaluate parameters for optimization}
\usage{
evaluatePars(p, pars_to_optimize, myfun, pars_bounds, fitness_measure, myfolder,
  key, otherPars)
}
\arguments{
\item{p}{numeric vector of evaluated values of parameters}

\item{pars_to_optimize}{names of arguments to \code{myfun} that should be
optimized}

\item{myfun}{the function being optimized: either 'segmentFolder' or
'analyzeFolder' (in quotes)}

\item{pars_bounds}{a list setting the lower and upper boundaries for possible
values of optimized parameters. For ex., if we optimize \code{smooth_ms}
and \code{smooth_overlap}, reasonable pars_bounds might be list(low = c(5,
0), high = c(500, 95))}

\item{fitness_measure}{the name of output variable that we are comparing with the key, e.g. 'nBursts' or 'pitch_median'}

\item{myfolder}{path to where the .wav files live}

\item{key}{a vector containing the "correct" measurement that we are aiming
to reproduce}
}
\value{
Returns 1 - Pearson's correlation between fitness measure and the key
  (i.e. 0 is perfect fit, 1 is awful fit).
}
\description{
Internal soundgen function.
}
\details{
Called by \code{\link{optimizePars}}.
}
