% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{findBursts}
\alias{findBursts}
\title{Find bursts}
\usage{
findBursts(envelope, timestep, interburst_min_ms, peak_to_global_max,
  peak_to_trough, trough_left = TRUE, trough_right = FALSE)
}
\arguments{
\item{envelope}{downsampled amplitude envelope}

\item{timestep}{time difference between two points in the envelope (ms)}

\item{interburst_min_ms}{minimum time between two consecutive bursts (ms). If
specified, it overrides \code{interburst_min_idx}}

\item{peak_to_global_max}{to qualify as a burst, a local maximum has to be at
least \code{peak_to_global_max} time the height of the global maximum of
the entire amplitude envelope}

\item{peak_to_trough}{to qualify as a burst, a local maximum has to be at
least \code{peak_to_trough}  times the local minimum on the LEFT over
analysis window (which is controlled by \code{interburst_min_ms} or
\code{interburst_min_scale})}

\item{trough_left}{should local maxima be compared to the trough
on the left and/or right of it? TRUE / FALSE}

\item{trough_right}{should local maxima be compared to the trough
on the left and/or right of it? TRUE / FALSE}
}
\value{
Returns a dataframe with timing of bursts
}
\description{
Internal soundgen function.
}
