% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{findGrad}
\alias{findGrad}
\title{Find gradient}
\usage{
findGrad(path, interpol = 3)
}
\arguments{
\item{path}{numeric vector corresponding to a path through pitch candidates}

\item{interpol}{the number of points to interpolate beyond each end of the path}
}
\value{
Returns a vector of the same length as input path giving its 4th derivative.
}
\description{
Internal soundgen function.
}
\details{
Internal helper function for postprocessing of pitch contour. Returns
gradient of internal energy of a snake. See \code{\link{snake}}.
}
