% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{findSyllables}
\alias{findSyllables}
\title{Find fyllables}
\usage{
findSyllables(envelope, timestep, threshold, shortest_syl, shortest_pause,
  merge_close_syl)
}
\arguments{
\item{envelope}{downsampled amplitude envelope}

\item{timestep}{time difference between two points in the envelope (ms)}

\item{threshold}{all continuous segments above this value are considered to
be syllables}

\item{shortest_syl}{minimum acceptable length of syllables (ms)}

\item{shortest_pause}{minimum acceptable break between syllables (ms).
Syllables separated by less time are merged. To avoid merging, specify
\code{shortest_pause = NA}}

\item{merge_close_syl}{if TRUE, syllable separated by less than
\code{shortest_pause} will be merged}
}
\value{
Returns a dataframe with timing of syllables.
}
\description{
Internal soundgen function.
}
