% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{findVoicedSegments}
\alias{findVoicedSegments}
\title{Find voiced segments}
\usage{
findVoicedSegments(pitchCands, shortest_syl = 20, shortest_pause = 50, step,
  samplingRate, min_voiced_cands = 2)
}
\arguments{
\item{pitchCands}{matrix of possible pitch values per column. One column is
one fft frame, one row is one pitch candidate}

\item{shortest_syl}{the smallest length of a voiced segment (ms) that
constitutes a syllable (shorter segments will be replaced by NA as if
unvoiced)}

\item{shortest_pause}{the smallest gap between voiced syllables (ms) that
means they shouldn't be merged into one voiced syllable}

\item{step}{fft step (ms)}

\item{samplingRate}{sampling rate (Hz)}

\item{min_voiced_cands}{a frame is considered to be voiced if at least this
many pitch candidates are not NA. Defaults to 2: since dom is usually
defined, in practice this means that we also want at least one other pitch
candidate (autocor, cep or BaNa)}
}
\value{
Returns a dataframe specifying where each voiced segment starts and
  ends (in fft frames, not ms!)
}
\description{
Internal soundgen function.
}
\details{
Internal helper function for postprocessing of pitch contours. Merges voiced
segments at least \code{shortest_syl} ms long and separated by less than
\code{shortest_pause} ms.
}
