% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{forcePerPath}
\alias{forcePerPath}
\title{Force per path}
\usage{
forcePerPath(pitch, pitchCands, pitchCert, pitchCenterGravity, certWeight)
}
\value{
Returns ???
}
\description{
Internal soundgen function.
}
\details{
Internal helper function for postprocessing of pitch contour. Returns the
total force acting on a snake (sum of internal and external gradients, i.e.
of the elastic force trying to straighten the snake [internal] and of the
force pushing the snake towards the most certain pitch estimates [external])
}
\examples{

}
