% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_contours.R
\name{getDiscreteContour}
\alias{getDiscreteContour}
\title{Discrete smooth contour from anchors}
\usage{
getDiscreteContour(len, anchors = data.frame(time = c(0, 1), ampl = c(1, 1)),
  method = c("spline", "loess")[2], ampl_floor = NULL,
  ampl_ceiling = NULL, ylim = NULL, plot = F, ...)
}
\arguments{
\item{len}{the number of syllables (equivalently, the length of generated contour)}

\item{anchors}{a dataframe of anchors with specified time and amplitude.
\code{achors$time} can be in ms (with len=NULL) or in arbitrary units, eg 0
to 1 (with duration determined by len, which must then be provided in ms).
So anchors$time is assumed to be in ms if len=NULL and relative if len is
specified. \code{anchors$ampl} can be on any scale.}

\item{method}{smoothing method: 'spline' or 'loess' (default). Consider
'spline' if you have a lot of anchors and not much patience, since it is
much faster than 'loess', but it may produce weird results when there are
only a few anchors}

\item{ampl_floor}{lower/upper bounds for the contour}

\item{ampl_ceiling}{lower/upper bounds for the contour}

\item{plot}{(boolean) produce a plot?}

\item{...}{other plotting options}
}
\value{
Numeric vector.
}
\description{
Internal soundgen function.
}
\details{
A discrete version of \code{\link{getSmoothContour}} with modified plotting. Intended for plotting variation in parameters across syllables.
}
