% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{getDom}
\alias{getDom}
\title{Get lowest dominant frequency band}
\usage{
getDom(frame, samplingRate, bin, dom_smoothing, dom_threshold, pitch_floor)
}
\arguments{
\item{frame}{the real part of the spectrum of a frame, as returned by
\code{\link[stats]{fft}}}

\item{samplingRate}{sampling rate (Hz)}

\item{bin}{the width of one bin in spectrogram, Hz}

\item{dom_smoothing}{the width of smoothing interval (Hz) for finding
the lowest spectral peak (lowest dominant frequency band)}

\item{dom_threshold}{(0 to 1) to find the lowest dominant frequency band, we
do short-term FFT and take the lowest frequency with amplitude at least
dom_threshold}

\item{pitch_floor}{absolute bounds for pitch candidates (Hz)}
}
\value{
Returns a list of $dom (NA or numeric) and $dom_array
  (either NULL or a dataframe of pitch candidates).
}
\description{
Internal soundgen function.
}
\details{
Calculate the lowest frequency band in the spectrum above pitch_floor whose
power exceeds a certain threshold.
}
