% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{getEntropy}
\alias{getEntropy}
\title{Normalized entropy}
\usage{
getEntropy(x)
}
\arguments{
\item{x}{vector of non-negative floats}
}
\value{
Returns entropy of a vector on a scale 0 to 1 (i.e., normalized so
  that entropy does not depend on the length of input vector). Zeroes are
  transformed into 1e-10, negative values produce NaN.
}
\description{
Internal soundgen function.
}
