% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{getFrameBank}
\alias{getFrameBank}
\title{Frame bank}
\usage{
getFrameBank(sound, samplingRate, windowLength_points, wn, step, zp,
  filter = NULL)
}
\arguments{
\item{sound}{numeric vector}

\item{samplingRate}{sampling rate, Hz}

\item{windowLength_points}{fft window length, in points}

\item{wn}{window type}

\item{step}{fft step, ms}

\item{zp}{zero padding, points}

\item{filter}{fft window filter (defaults to NULL)}
}
\value{
A matrix with \code{nrow = windowLength_points/2} and \code{ncol}
  depending on \code{length(sound)} and \code{step}
}
\description{
Internal soundgen function.
}
\details{
A subroutine of \code{\link{spec}} that saves windowed (and
optionally zero-padded) frames, i.e. chunks of the sound file of the right
size and spacing. Handy for further processing.
}
\examples{
a = soundgen:::getFrameBank(sin(1:1000), 16000, 512, 'gaussian', 15, 0)
}
