% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{getMelSpec}
\alias{getMelSpec}
\title{Mel-transformed spectrum}
\usage{
getMelSpec(s, samplingRate = NULL, windowLength = 40, overlap = 50,
  step = NULL, throwaway = -120, plot = FALSE)
}
\arguments{
\item{plot}{if TRUE, plots the spectrum}
}
\description{
Internal soundgen function
}
\details{
Takes a .wav file or a waveform as numeric vector + samplingRate and returns
mel-transformed spectrum (auditory spectrum). See \code{\link{match}}.
}
