% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{getPitchSpec}
\alias{getPitchSpec}
\title{BaNa pitch tracker}
\usage{
getPitchSpec(frame, spec_smoothing, spec_peak_HNRslope, bin, HNR = NULL,
  spec_threshold, spec_peak, spec_singlePeakCert, pitch_floor, pitch_ceiling,
  spec_merge, nCands)
}
\arguments{
\item{frame}{the real part of the spectrum of a frame, as returned by
\code{\link[stats]{fft}}}

\item{spec_smoothing}{the width of window for detecting peaks in the spectrum, Hz}

\item{spec_peak_HNRslope}{when looking for putative
harmonics in the spectrum, the threshold for peak detection is calculated as
\code{spec_peak * (1 - HNR * spec_peak_HNRslope)}. For noisy sounds the
threshold is high to avoid false sumharmonics, while for tonal sounds it is low
to catch weak harmonics}

\item{bin}{the width of spectral bin in \code{frame}, Hz}

\item{HNR}{harmonics-to-noise ratio returned by \code{\link{getPitchAutocor}}}

\item{spec_threshold}{(0 to 1) separate
voicing thresholds for detecting pitch candidates with three different
methods: autocorrelation, cepstrum, and BaNa algorithm (see Details). Note
that HNR is still calculated for frames considered to be unvoiced.}

\item{spec_peak}{when looking for putative
harmonics in the spectrum, the threshold for peak detection is calculated as
\code{spec_peak * (1 - HNR * spec_peak_HNRslope)}. For noisy sounds the
threshold is high to avoid false sumharmonics, while for tonal sounds it is low
to catch weak harmonics}

\item{spec_singlePeakCert}{(0 to 1) if only one pitchSpec candidate is
found, its weight (certainty) is taken to be
\code{spec_singlePeakCert}}

\item{pitch_floor}{absolute bounds for pitch candidates (Hz)}

\item{pitch_ceiling}{absolute bounds for pitch candidates (Hz)}

\item{spec_merge}{pitch candidates within \code{spec_merge} are
merged with boosted certainty}

\item{nCands}{maximum number of pitch candidates to use per method (except
for \code{dom}, which returns at most one candidate per frame)}
}
\value{
Returns either NULL or a dataframe of pitch candidates.
}
\description{
Internal soundgen function.
}
\details{
Attempts to find F0 of a frame by calculating ratios of putative harmonics
(frequency domain analysis, ~ modified BaNa algorithm). See Ba et al. (2012)
"BaNa: A hybrid approach for noise resilient pitch detection." Statistical
Signal Processing Workshop (SSP), 2012 IEEE.
}
