% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shiny.R
\name{HzToSemitones}
\alias{HzToSemitones}
\title{Convert Hz to semitones}
\usage{
HzToSemitones(h)
}
\arguments{
\item{h}{vector of frequencies (Hz)}
}
\description{
Internal soundgen function.
}
\details{
Converts from Hz to semitones above C0 (~16.4 Hz).
}
