% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_segment.R
\name{mergeSyllables}
\alias{mergeSyllables}
\title{Merge syllables}
\usage{
mergeSyllables(syllables, shortest_pause)
}
\arguments{
\item{syllables}{a dataframe listing syllables with time_start and time_end}

\item{shortest_pause}{minimum acceptable break between syllables (ms).
Syllables separated by less time are merged. To avoid merging, specify
\code{shortest_pause = NA}}
}
\description{
Internal soundgen function.
}
\details{
Merges syllables if they are separated by less than \code{shortest_pause ms}. Called by \code{\link{findSyllables}}.
}
