% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{morph}
\alias{morph}
\title{Morph sounds}
\usage{
morph(formula1, formula2, nHybrids, playMorphs = TRUE, savePath = NA,
  samplingRate = 16000)
}
\arguments{
\item{formula1, formula2}{lists of parameters for calling \code{\link{soundgen}} that produce the two target sounds between which morphing will occur}

\item{nHybrids}{the length of morphing sequence, including target sounds}

\item{playMorphs}{if TRUE, the morphing sequence will be played}

\item{savePath}{if it is the path to an existing directory, morphs will be saved there as individual .wav files (defaults to NA)}

\item{samplingRate}{sampling rate of output (Hz)}
}
\value{
A list of two sublists ('formulas' and 'sounds'), each sublist of length nHybrids. For ex., the formula for the second hybrid is m$formulas[[2]], and the waveform is m$sounds[[2]]
}
\description{
Takes two formulas for synthesizing two target sounds and produces a number of intermediate forms (morphs), attempting to go from one target sound to the other in a specified number of equal steps.
}
\examples{
# write or copy-paste two formulas from the app soundgen_app(), for example:
formula1 = list(sylLen=1190, rolloff=-24, rolloffAdjust_quadratic=20,
  rolloffAdjust_quadratic_nHarm=1, rolloff_noise=-13,
  subFreq=100, subDep=100,
  pitchAnchors=data.frame(time=c(0,.05,.25,.35,1), value=c(240,305,250,260,250)),
  noiseAnchors=data.frame(time=c(-50,1214), value=c(-50,-70)),
  mouthAnchors=data.frame(time=c(0,0.1,.4,1), value=c(0.18,0.5,.2,.1)),
  amplAnchors=data.frame(time=c(0,0.16,1), value=c(19,120,120)),
  exactFormants=list(f1=data.frame(time=c(0), freq=c(700),
  amp=c(30), width=c(80)), f2=data.frame(time=c(0), freq=c(900), amp=c(30),
  width=c(120)), f3=data.frame(time=c(0), freq=c(1500), amp=c(20),
  width=c(150))))
formula2 = list(sylLen=240, pitchEffects_amount=100,
  pitchEffects_intensity=50, attackLen=1, jitterDep=5.1, jitterLen=1,
  formantDep=1.3,
  extraFormants_stochastic=15, rolloff=-12,
  rolloffAdjust_per_octave=-6, rolloffAdjust_per_kHz=-6,
  rolloff_noise=-13, temperature=0.01,
  pitchAnchors=data.frame(time=c(0,0.17,1), value=c(383,421,358)),
  noiseAnchors=data.frame(time=c(0,49,256), value=c(-120,10,-120)),
  mouthAnchors=data.frame(time=c(0,0.15,1), value=c(0,0.49,0.2)),
  amplAnchors=data.frame(time=c(0,0.26,1), value=c(43,120,43)),
  exactFormants=list(f1=data.frame(time=c(0), freq=c(400), amp=c(40),
  width=c(120)), f2=data.frame(time=c(0), freq=c(1000), amp=c(40),
  width=c(120)), f3=data.frame(time=c(0), freq=c(1500), amp=c(30),
  width=c(150))))

 m = morph (formula1, formula2, nHybrids=5, playMorphs=TRUE)
 # use m$formulas to access formulas for each morph, m$sounds for waveforms
 playme(m$sounds[[2]], 16000)
}
