% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_morph.R
\name{morphFormants}
\alias{morphFormants}
\title{Morph formants}
\usage{
morphFormants(f1, f2, nHybrids = 5)
}
\arguments{
\item{f1, f2}{dataframes specifying one formant of the two target sounds
(different numbers of rows are ok)}

\item{nHybrids}{length of morphing sequence}
}
\description{
Internal soundgen function.
}
