% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{pathfinder}
\alias{pathfinder}
\title{Pathfinder}
\usage{
pathfinder(pitchCands, pitchCert, certWeight = 0.5, interpolWindow = 3,
  interpolTolerance = 0.05, interpolCert = 0.3, runSnake = T,
  snakeSmoothingStep = 0.05, snakeIterMultiplier = 2, plotSnake = F)
}
\description{
Internal soundgen function.
}
\details{
Internal helper function for postprocessing pitch contour. Starts with a
reasonable guess and computes the more-or-less optimal pitch contour (not
quite the very optimal - too computationally expensive).
}
\examples{

}
