% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shiny.R
\name{pickle}
\alias{pickle}
\title{Convert list of formants to string}
\usage{
pickle(l)
}
\arguments{
\item{l}{a list of exact formants}
}
\description{
Internal soundgen function.
}
\details{
Helper function that converts a list of exact formants into a human-readable
string, which can then be modified and evaluated to reproduce a list of the
original structure. Note: dput() and dump() can pickle & reproduce any
object, but I can't figure out how to make the interim string human-readable
and modifiable
}
