% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{saveme}
\alias{saveme}
\title{Save wave file}
\usage{
saveme(sound, filename, samplingRate = 44100)
}
\arguments{
\item{sound}{numeric vector on any scale}

\item{filename}{path to target .wav file}

\item{samplingRate}{sampling rate of target .wav file}
}
\description{
Internal soundgen function.
}
\details{
Saves a numeric vector as a .wav file. Provides a convenient wrapper for the
functionality of tuneR package.
}
\examples{
\dontrun{saveme(sin(2*pi*f0_Hz*(1:16000)/16000), '~/Downloads/temp.wav', samplingRate=16000)}
}
