% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSM.R
\name{selfsim}
\alias{selfsim}
\title{Compute self-similarity}
\usage{
selfsim(m, norm = FALSE, simil = c("cosine", "correlation")[1], win = 1)
}
\arguments{
\item{m}{input matrix such as a spectrogram}

\item{norm}{if TRUE, each fft frame is normalized by max power. NB:
recommended if using cosine similarity with spectrum (audiogram)}

\item{simil}{method for comparing frames: either cosine similarity or
Pearson's correlation}

\item{win}{window for averaging SSM (reduces its size). Forced to de odd}
}
\value{
Returns a square self-similarity matrix.
}
\description{
Internal soundgen function.
}
\details{
Called by \code{\link{ssm}}.
}
