% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{semitonesToHz}
\alias{semitonesToHz}
\title{Convert semitones to Hz}
\usage{
semitonesToHz(s)
}
\arguments{
\item{s}{vector of frequencies (semitones above C0)}
}
\description{
Internal soundgen function.
}
\details{
Converts from semitones above C0 (~16.4 Hz) to Hz
}
