% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundgen.R
\docType{package}
\name{soundgen}
\alias{soundgen}
\alias{soundgen-package}
\title{soundgen: A package for parametric synthesis of voice and acoustic analysis.}
\description{
The soundgen package provides ? categories of important functions:
generateBout, bla-bla and bla-bla.
}
