% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{to_dB}
\alias{to_dB}
\title{Convert to dB}
\usage{
to_dB(x)
}
\arguments{
\item{x}{a vector of floats between 0 and 1 (exclusive, i.e. these are ratios)}
}
\description{
Internal soundgen function.
}
