% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_contours.R
\name{wiggleAnchors}
\alias{wiggleAnchors}
\title{Randomly modify anchors}
\usage{
wiggleAnchors(df, temperature, temp_coef, low, high, wiggleAllRows = F)
}
\arguments{
\item{df}{dataframe of anchors, for ex. \code{data.frame(time = c(0, .1, .8,
1), ampl = c(100, 230, 180, 90))}}

\item{temperature, temp_coef}{regulate the amount of stochasticity
("wiggling"). Since \code{temperature} is used in several functions,
\code{temp_coef} gives more flexibility by controlling how much temperature
affects this particular aspect, namely random variation in anchors. These
two are multiplied, so \code{temp_coef} of 0.5 halves the effect of
temperature.}

\item{low, high}{bounds on possible variation. Both \code{low} and \code{high}
should be vectors of length 2: the first element specifies the boundary for
\code{df$time} and the second for \code{df$ampl}. Ex.: low = c(0,1) - low
bound on "time"=0, low bound on "ampl"=1}

\item{wiggleAllRows}{should the firts and last time anchors be wiggled? (TRUE
for breathing, FALSE for other anchors)}
}
\value{
Modified original dataframe.
}
\description{
Internal soundgen function.
}
\details{
A helper function for introducing random variation into any anchors (for
pitch / breathing / amplitude / ...). NB: if the lower and upper bounds are
unreasonable given the scale of df$ampl, \code{wiggleAnchors} will keep
running forever!
}
\examples{
wiggleAnchors(df = data.frame(time = c(0, .1, .8, 1), ampl = c(100, 230, 180, 90)),
  temperature=.2, temp_coef=.1, low=c(0, 50), high=c(1, 1000), wiggleAllRows = F) # pitch
wiggleAnchors(df = data.frame(time = c(-100, 100, 600, 900), ampl = c(-120, -80, 0, -120)),
  temperature=.1, temp_coef=.5, low=c(-Inf, -120), high=c(+Inf, 30), wiggleAllRows = T) # breathing
}
