% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{findGrad}
\alias{findGrad}
\title{Find gradient}
\usage{
findGrad(path, interpol = 3)
}
\value{
Returns ???
}
\description{
Internal soundgen function.
}
\details{
Internal helper function for postprocessing of pitch contour. Returns
gradient of internal energy of a snake. See \code{\link{snake}}.
}
