% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{getEntropy}
\alias{getEntropy}
\title{Shannon entropy}
\usage{
getEntropy(x)
}
\arguments{
\item{x}{vector of non-negative floats}
}
\value{
Float or NA
}
\description{
Internal soundgen function.
}
\details{
Returns Shannon entropy of a vector. Zeroes are dealt with by adding 1 to all
elements. If all elements are zero, returns NA.
}
