% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSM.R
\name{getNovelty}
\alias{getNovelty}
\title{SSM novelty}
\usage{
getNovelty(ssm, kernel_size, kernel_sd)
}
\arguments{
\item{ssm}{self-similarity matrix, as produced by \code{\link{selfsim}}}

\item{kernel_size}{the size of gaussian kernel (points)}

\item{kernel_sd}{the SD of gaussian kernel}
}
\value{
Returns a numeric vector of length \code{nrow(ssm)}
}
\description{
Internal soundgen function.
}
\details{
Calculates novelty in a self-similarity matrix. See El Badawy, D., Marmaroli,
P., & Lissek, H. (2013). Audio Novelty-Based Segmentation of Music Concerts.
In Acoustics 2013 (No. EPFL-CONF-190844).
}
