% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{getPitchCep}
\alias{getPitchCep}
\title{Cepstral pitch tracker}
\usage{
getPitchCep(frame, cep_zp, samplingRate, pitch_floor, pitch_ceiling,
  cep_threshold, cep_smoothing = NULL, nCands)
}
\arguments{
\item{frame}{the real part of the spectrum of a frame, as returned by
\code{\link[stats]{fft}}}

\item{cep_zp}{zero-padding of the spectrum used for cepstral pitch detection
(points). Improves the precision of cepstral pitch detection quite
noticeably.}

\item{samplingRate}{sampling rate (Hz)}

\item{pitch_floor}{absolute bounds for pitch candidates (Hz)}

\item{pitch_ceiling}{absolute bounds for pitch candidates (Hz)}

\item{cep_threshold}{(0 to 1) separate
voicing thresholds for detecting pitch candidates with three different
methods: autocorrelation, cepstrum, and BaNa algorithm (see Details). Note
that HNR is still calculated for frames considered to be unvoiced.}

\item{cep_smoothing}{the width of smoothing interval (in bins) for finding
peaks in the cepstrum. Defaults to 7 for sampling rate 44100 and smaller
odd numbers for lower values of sampling rate}

\item{nCands}{maximum number of pitch candidates to use per method (except
for \code{dom}, which returns at most one candidate per frame)}
}
\value{
Returns either NULL or a dataframe of pitch candidates.
}
\description{
Internal soundgen function.
}
\details{
Attempts to find F0 of a frame by looking for peaks in the cepstrum.
See http://www.phon.ucl.ac.uk/courses/spsci/matlab/lect10.html
}
