% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{getPitchSpec}
\alias{getPitchSpec}
\title{BaNa pitch tracker}
\usage{
getPitchSpec(frame, width_spec, slope_spec, bin, HNR = NULL,
  voiced_threshold_spec, specPitchThreshold_nullNA, pitchSpec_only_peak_weight,
  pitch_floor, pitch_ceiling, merge_semitones, nCands)
}
\arguments{
\item{frame}{the real part of the spectrum of a frame, as returned by
\code{\link[stats]{fft}}}

\item{width_spec}{the width of window for detecting peaks in the spectrum
(BaNa - spectral pitch tracking)}

\item{slope_spec}{when looking for putative
harmonics in the spectrum, the threshold for peak detection is calculated as
\code{specPitchThreshold_nullNA * (1 - HNR * slope_spec)}. For noisy sounds the
threshold is high to avoid false sumharmonics, while for tonal sounds it is low
to catch weak harmonics (BaNa - spectral pitch tracking)}

\item{bin}{the width of spectral bin in \code{frame}, Hz}

\item{HNR}{harmonics-to-noise ratio returned by \code{\link{getPitchAutocor}}}

\item{voiced_threshold_spec}{(0
to 1) separate thresholds for detecting pitch candidates with three
different methods: autocorrelation, cepstrum, and BaNa algorithm (see
Details). Note that HNR is still calculated for frames considered to be
unvoiced.}

\item{specPitchThreshold_nullNA}{when looking for putative
harmonics in the spectrum, the threshold for peak detection is calculated as
\code{specPitchThreshold_nullNA * (1 - HNR * slope_spec)}. For noisy sounds the
threshold is high to avoid false sumharmonics, while for tonal sounds it is low
to catch weak harmonics (BaNa - spectral pitch tracking)}

\item{pitchSpec_only_peak_weight}{(0 to 1) if only one pitchSpec candidate is
found, its weight (certainty) is taken to be
\code{pitchSpec_only_peak_weight}. This mainly has implications for how
much we trust the BaNa estimate vs. the autocorrelation estimate of f0.}

\item{pitch_floor}{bounds for pitch candidates (Hz)}

\item{pitch_ceiling}{bounds for pitch candidates (Hz)}

\item{merge_semitones}{pitch candidates within \code{merge_semitones} are
merged with boosted certainty (BaNa - spectral pitch tracking)}

\item{nCands}{maximum number of pitch candidates to use per method (dom and
cepstrum always use only one candidate each).}
}
\value{
Returns either NULL or a dataframe of pitch candidates.
}
\description{
Internal soundgen function.
}
\details{
Attempts to find F0 of a frame by calculating ratios of putative harmonics
(frequency domain analysis, ~ modified BaNa algorithm). See Ba et al. (2012)
"BaNa: A hybrid approach for noise resilient pitch detection." Statistical
Signal Processing Workshop (SSP), 2012 IEEE.
}
