% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSM.R
\name{selfsim}
\alias{selfsim}
\title{Compute self-similarity}
\usage{
selfsim(m, normalizeInput = FALSE, simil_measure = c("cosine",
  "correlation")[1], win = 1)
}
\arguments{
\item{m}{input matrix such as a spectrogram}

\item{normalizeInput}{if TRUE, each fft frame is normalized by max power. NB: do this if using cosine similarity with spectrum (audiogram)}

\item{simil_measure}{method for comparing frames: either cosine similarity or Pearson's correlation}

\item{win}{window for averaging SSM (reduces its size). Forced to odd: 1 3 5 etc}
}
\value{
Returns a square self-similarity matrix.
}
\description{
Internal soundgen function.
}
