% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{snake}
\alias{snake}
\title{Snake}
\usage{
snake(pitch, pitchCands, pitchCert, certWeight, snakeSmoothingStep = 0.05,
  snakeIterMultiplier = 2, plotSnake = F)
}
\description{
Internal soundgen function.
}
\details{
Internal helper function for postprocessing of pitch contour. Wiggles a snake along the gradient of internal + external forces. NB: if the snake is run, the final contour may deviate from the actually measured pitch candidates!
}
\examples{

}
