% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSM.R
\name{ssm}
\alias{ssm}
\title{Self-similarity matrix}
\usage{
ssm(x, samplingRate = NULL, windowLength = 40, step = NULL,
  maxFreq = NULL, nBands = NULL, MFCC = 2:13, input = c("cepstrum",
  "audiogram")[1], norm = FALSE, simil = c("cosine", "correlation")[1],
  returnSSM = TRUE, kernelLen = 200, kernelSD = 0.2, win = 3,
  plot = TRUE, specPars = list(levels = seq(0, 1, length = 30),
  color.palette = seewave::spectro.colors, xlab = "Time, ms", ylab = "kHz", ylim
  = c(0, maxFreq/1000)), ssmPars = list(levels = seq(0, 1, length = 30),
  color.palette = seewave::spectro.colors, xlab = "Time, ms", ylab = "Time, ms",
  main = "Self-similarity matrix"), noveltyPars = list(type = "b", pch = 16,
  col = "black", lwd = 3))
}
\arguments{
\item{x}{path to a .wav file or a vector of amplitudes with specified
samplingRate}

\item{samplingRate}{sampling rate of \code{x} (only needed if \code{x} is a
numeric vector, rather than a .wav file)}

\item{windowLength}{length of fft window (ms)}

\item{step}{fft step (ms). Defaults to \code{windowLength / 4}}

\item{maxFreq}{highest band edge of mel filters (Hz). Defaults to
\code{samplingRate / 2}. See \code{\link[tuneR]{melfcc}}}

\item{nBands}{number of warped spectral bands to use. Defaults to \code{100 *
windowLength / 20}. See \code{\link[tuneR]{melfcc}}}

\item{MFCC}{if \code{input = 'audiogram'}, which mel-frequency cepstral
coefficients to use. Ddefaults to \code{2:13})}

\item{input}{either MFCCs ("cepstrum") or mel-filtered spectrum ("audiogram")}

\item{norm}{if TRUE, each fft frame is normalized by max power. NB:
recommended if using cosine similarity with spectrum (audiogram)}

\item{simil}{method for comparing frames: either cosine similarity or
Pearson's correlation}

\item{returnSSM}{if TRUE, returns the SSM}

\item{kernelLen}{length of checkerboard kernel for calculating novelty, ms (the
larger, the more global vs. local the novelty)}

\item{kernelSD}{SD of checkerboard kernel for calculating novelty}

\item{win}{window for averaging SSM (reduces its size). Forced to de odd}

\item{plot}{if TRUE, plots the SSM}

\item{specPars}{graphical parameters passed to
\code{seewave::filled.contour.modif2} and affecting the spectrogram}

\item{ssmPars}{graphical parameters passed to
\code{seewave::filled.contour.modif2} and affecting the plot of SSM}

\item{noveltyPars}{graphical parameters passed to
\code{\link[graphics]{lines}} and affecting the novelty contour}
}
\value{
If \code{returnSSM} is TRUE, returns a list of two components: $ssm
  contains the self-similarity matrix, and $novelty contains the novelty
  vector. If \code{returnSSM} is FALSE, only produces a plot.
}
\description{
Calculates the self-similarity matrix and novelty vector for a sound.
}
\examples{
sound = c(soundgen(), soundgen(nSyl = 4, sylLen = 200,
  pauseLen = 140,  pitchAnchors = list(time = c(0, 0.65, 1),
  value = c(977, 1540, 826)), formants = NA))
s = ssm(sound, samplingRate = 16000,
        input = 'cepstrum', win = 3, kernelLen = 350)  # detailed, local features
s = ssm(sound, samplingRate = 16000,
        input = 'cepstrum', win = 7, kernelLen = 800)  # more global
plot(s$novelty, type='b')  # use for peak detection, etc
}
\references{
\itemize{
  \item El Badawy, D., Marmaroli, P., & Lissek, H. (2013). Audio
  Novelty-Based Segmentation of Music Concerts. In Acoustics 2013 (No.
  EPFL-CONF-190844)
  \item Foote, J. (1999, October). Visualizing music and
  audio using self-similarity. In Proceedings of the seventh ACM
  international conference on Multimedia (Part 1) (pp. 77-80). ACM.
  \item
  Foote, J. (2000). Automatic audio segmentation using a measure of audio
  novelty. In Multimedia and Expo, 2000. ICME 2000. 2000 IEEE International
  Conference on (Vol. 1, pp. 452-455). IEEE.
  }
}
